building-agent/requirements.txt
openai>=1.40
python-dotenv>=1.0
httpx>=0.27
